Forms: Audit & normalize form-field sanitization.
* Make sure they all work the same
* Use filters in `core/filters.php` where appropriate
* Add `absint` filters for numeric values
* Add filters for per-forum moderator text field
* Remove any accidental duplicate escaping

git-svn-id: https://bbpress.svn.wordpress.org/trunk@6248 9866e705-20ec-0310-96e7-cbb4277adcfb

+++ src/includes/replies/template.php
@@ -2600,7 +2600,7 @@ function bbp_form_reply_to() {
 			$reply_to = bbp_get_reply_to();
 		}
 
-		return (int) apply_filters( 'bbp_get_form_reply_to', $reply_to );
+		return apply_filters( 'bbp_get_form_reply_to', $reply_to );
 	}
 
 /**
